+++ trelby/autocompletion.py
@@ -1,8 +1,7 @@
-import trelby.config as config
 import trelby.mypickle as mypickle
 import trelby.screenplay as screenplay
 import trelby.util as util
-
+from trelby.autocompletiontype import AutoCompletionType as Type
 
 # manages auto completion information for a single script.
 class AutoCompletion:
@@ -75,33 +74,3 @@ class AutoCompletion:
     # get type's Type, or None if it doesn't exist.
     def getType(self, lt):
         return self.types.get(lt)
-
-
-# auto completion info for one element type
-class Type:
-    cvars = None
-
-    def __init__(self, lt):
-
-        # pointer to TypeInfo
-        self.ti = config.lt2ti(lt)
-
-        if not self.__class__.cvars:
-            v = self.__class__.cvars = mypickle.Vars()
-
-            v.addBool("enabled", True, "Enabled")
-            v.addList("items", [], "Items", mypickle.StrLatin1Var("", "", ""))
-
-            v.makeDicts()
-
-        self.__class__.cvars.setDefaults(self)
-
-    def save(self, prefix):
-        prefix += "%s/" % self.ti.name
-
-        return self.cvars.save(prefix, self)
-
-    def load(self, vals, prefix):
-        prefix += "%s/" % self.ti.name
-
-        self.cvars.load(vals, prefix, self)
