+++ trelby/autocompletiontype.py
@@ -1,0 +1,31 @@
+import trelby.config as config
+import trelby.mypickle as mypickle
+
+# auto completion info for one element type
+class AutoCompletionType:
+    cvars = None
+
+    def __init__(self, lt):
+
+        # pointer to TypeInfo
+        self.ti = config.lt2ti(lt)
+
+        if not self.__class__.cvars:
+            v = self.__class__.cvars = mypickle.Vars()
+
+            v.addBool("enabled", True, "Enabled")
+            v.addList("items", [], "Items", mypickle.StrLatin1Var("", "", ""))
+
+            v.makeDicts()
+
+        self.__class__.cvars.setDefaults(self)
+
+    def save(self, prefix):
+        prefix += "%s/" % self.ti.name
+
+        return self.cvars.save(prefix, self)
+
+    def load(self, vals, prefix):
+        prefix += "%s/" % self.ti.name
+
+        self.cvars.load(vals, prefix, self)
